refactor: organize notification components
