+++ frontend/src/subscription.js
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { siteRoot, mediaUrl, logoPath, logoWidth, logoHeight, siteTitle } from './utils/constants';
 import SideNav from './components/user-settings/side-nav';
 import Account from './components/common/account';
-import Notification from './components/common/notification';
+import Notification from './components/notification';
 import Subscription from './components/subscription';
 
 import './css/toolbar.css';
